Split page path utils
