+++ packages/next/server/dev/on-demand-entry-handler.ts
@@ -5,8 +5,10 @@ import { EventEmitter } from 'events'
 import { findPageFile } from '../lib/find-page-file'
 import { getPageRuntime, runDependingOnPageType } from '../../build/entries'
 import { join, posix } from 'path'
-import { normalizePagePath, normalizePathSep } from '../page-path-utils'
-import { ensureLeadingSlash, removePagePathTail } from '../page-path-utils'
+import { normalizePathSep } from '../../shared/lib/page-path/normalize-path-sep'
+import { normalizePagePath } from '../../shared/lib/page-path/normalize-page-path'
+import { ensureLeadingSlash } from '../../shared/lib/page-path/ensure-leading-slash'
+import { removePagePathTail } from '../../shared/lib/page-path/remove-page-path-tail'
 import { pageNotFoundError } from '../require'
 import { reportTrigger } from '../../build/output'
 import getRouteFromEntrypoint from '../get-route-from-entrypoint'
