+++ packages/next/server/require.ts
@@ -7,8 +7,9 @@ import {
   SERVER_DIRECTORY,
   SERVERLESS_DIRECTORY,
 } from '../shared/lib/constants'
-import { normalizePagePath, denormalizePagePath } from './page-path-utils'
 import { normalizeLocalePath } from '../shared/lib/i18n/normalize-locale-path'
+import { normalizePagePath } from '../shared/lib/page-path/normalize-page-path'
+import { denormalizePagePath } from '../shared/lib/page-path/denormalize-page-path'
 import type { PagesManifest } from '../build/webpack/plugins/pages-manifest-plugin'
 import type { MiddlewareManifest } from '../build/webpack/plugins/middleware-plugin'
 import type { WasmBinding } from '../build/webpack/loaders/get-module-build-info'
