+++ packages/next/shared/lib/i18n/get-locale-metadata.ts
@@ -1,5 +1,5 @@
 import { acceptLanguage } from '../../../server/accept-header'
-import { denormalizePagePath } from '../../../server/page-path-utils'
+import { denormalizePagePath } from '../page-path/denormalize-page-path'
 import { detectDomainLocale } from './detect-domain-locale'
 import { formatUrl } from '../router/utils/format-url'
 import { normalizeLocalePath } from './normalize-locale-path'
