+++ test/unit/isolated/require-page.test.ts
@@ -1,8 +1,8 @@
 /* eslint-env jest */
 
 import { join } from 'path'
-import { normalizePagePath } from 'next/dist/server/page-path-utils'
 import { SERVER_DIRECTORY, CLIENT_STATIC_FILES_PATH } from 'next/constants'
+import { normalizePagePath } from 'next/dist/shared/lib/page-path/normalize-page-path'
 import {
   requirePage,
   getPagePath,
